pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptr"
version = "0.1.0"
description = "Promptr — a GNOME teleprompter app built with GTK4 and libadwaita"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Headless Flower" }
]
license = { text = "MIT" }
dependencies = [
  "python-docx>=1.1.0",
  "pypdf>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.11.0",
  "mypy>=1.10.0",
]

[project.scripts]
promptr = "promptr.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
